+++ src/main/java/hk/org/ha/kcc/its/repository/ServiceRepository.java
@@ -3,5 +3,5 @@ package hk.org.ha.kcc.its.repository;
 import hk.org.ha.kcc.its.model.Services;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ServiceRepository extends JpaRepository<Services, Long> {
+public interface ServiceRepository extends JpaRepository<Services, Integer> {
 }
